refactor: 개행 일관성 및 상수처리
