+++ backend/src/main/java/com/company/dinner/auth/config/AuthConfig.java
@@ -14,11 +14,8 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import java.util.List;
 
 import static com.company.dinner.auth.controller.interceptor.HttpMethod.ANY;
-import static com.company.dinner.auth.controller.interceptor.HttpMethod.DELETE;
 import static com.company.dinner.auth.controller.interceptor.HttpMethod.GET;
 import static com.company.dinner.auth.controller.interceptor.HttpMethod.OPTIONS;
-import static com.company.dinner.auth.controller.interceptor.HttpMethod.PATCH;
-import static com.company.dinner.auth.controller.interceptor.HttpMethod.POST;
 
 @RequiredArgsConstructor
 @Configuration
@@ -37,21 +34,12 @@ public class AuthConfig implements WebMvcConfigurer {
     private HandlerInterceptor loginCheckerInterceptor() {
         return new PathMatcherInterceptor(loginCheckerInterceptor)
                 .excludePathPattern("/**", OPTIONS)
-                .addPathPatterns("/stations/**/reports/me", GET);
+                .addPathPatterns("/tests/**", GET);
     }
 
     private HandlerInterceptor loginInterceptor() {
         return new PathMatcherInterceptor(loginInterceptor)
-                .excludePathPattern("/**", OPTIONS)
-                .addPathPatterns("/admin/**", ANY)
-                .addPathPatterns("/members/**", ANY)
-                .addPathPatterns("/cars/**/filters", POST)
-                .addPathPatterns("/cars", POST)
-                .addPathPatterns("/filters", POST)
-                .addPathPatterns("/stations/**/reports", POST)
-                .addPathPatterns("/stations/**/reviews/**", POST, DELETE, PATCH)
-                .addPathPatterns("/reviews/**/replies", POST, DELETE, PATCH)
-                .addPathPatterns("/reviews/**", DELETE, PATCH);
+                .excludePathPattern("/**", ANY, OPTIONS);
     }
 
     @Override
