+++ backend/src/main/java/com/company/dinner/auth/controller/support/AuthenticationExtractor.java
@@ -9,6 +9,10 @@ public class AuthenticationExtractor {
 
     private static final String AUTHORIZATION_HEADER = "Authorization";
     private static final String BEARER = "Bearer";
+    private static final String HEADER_SPLIT_DELIMITER = " ";
+    private static final int TOKEN_TYPE_INDEX = 0;
+    private static final int TOKEN_VALUE_INDEX = 1;
+    private static final int VALID_HEADER_SPLIT_LENGTH = 2;
 
     public static Optional<String> extract(final HttpServletRequest request) {
         String header = request.getHeader(AUTHORIZATION_HEADER);
@@ -17,12 +21,13 @@ public class AuthenticationExtractor {
             return Optional.empty();
         }
 
-        return extractFromHeader(header.split(" "));
+        return extractFromHeader(header.split(HEADER_SPLIT_DELIMITER));
     }
 
     public static Optional<String> extractFromHeader(final String[] headerParts) {
-        if (headerParts.length == 2 && headerParts[0].equals(BEARER)) {
-            return Optional.ofNullable(headerParts[1]);
+        if (headerParts.length == VALID_HEADER_SPLIT_LENGTH &&
+                headerParts[TOKEN_TYPE_INDEX].equals(BEARER)) {
+            return Optional.ofNullable(headerParts[TOKEN_VALUE_INDEX]);
         }
 
         return Optional.empty();
