+++ backend/src/main/java/com/company/dinner/member/domain/Member.java
@@ -18,13 +18,11 @@ import javax.persistence.Id;
 
 @Getter
 @Builder
+@EqualsAndHashCode(of = "id", callSuper = false)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@EqualsAndHashCode(of = "id", callSuper = false)
 @Entity
 public class Member extends BaseEntity {
-
-    private static final int EMAIL_MASKING_LENGTH = 2;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
